Fixed pandoc inivisible selection
When selecting code, a small mysterious box would appear a tabs space 
away from the end of each line. This was caused by the span class, with 
the CSS property "left:-4em". When selecting a span class, it will 
select a space after the span even if there is nothing there. This space 
doesn't actually move when the "left" property is used, resulting in a 
disjointed selection. By replacing that with "margin-left:-4em", this 
bug was resolved.